pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "agentic"
version = "0.1.0"
description = "A Python project for agentic systems"
authors = [{ name = "LatentVectors" }]

readme = "README.md"
requires-python = ">=3.12"
classifiers = ["Programming Language :: Python :: 3.12"]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.3.0",
    "langchain-chroma>=0.1.0",
    "langgraph>=0.2.0",
    "chromadb>=0.6.0",
    "beautifulsoup4>=4.12.0",
    "loguru>=0.7.0",
    "typer>=0.9.0",
    "tqdm>=4.65.0",
    "grandalf>=0.1.0",
    "rich>=13.0.0",
    "openevals>=0.1.0",
    "vcrpy>=1.0.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.1.0", "mypy>=1.5.0", "pytest>=7.0.0"]

[project.scripts]
agentic = "src.cli:app"

[tool.setuptools.packages.find]
where = ["src", "."]

[tool.ruff]
line-length = 99
src = ["src"]
extend-exclude = [".venv", "data", "logs", "build", "dist", "__pycache__"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
extend-select = ["I"]


[tool.mypy]
files = ["src/**/*.py", "test/**/*.py"]
exclude = "(build|dist|\\.venv|data|logs|__pycache__)/"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
